requirements.txt
# Core packages
streamlit==1.35.0
langchain>=0.1.15
langchain-community>=0.0.36
langchain-openai>=0.1.6
openai>=1.30.0

# Utilities
pydantic>=2.0,<3.0
faiss-cpu==1.7.4
PyMuPDF==1.23.21
fitz
pandas>=2.0.3
numpy>=1.26.4
sentence_transformers>=2.2.2
torch>=2.0.1
altair>=4.2.0
plotly>=5.15.0
scikit-learn>=1.3.0
plotly.express>=0.4.1
hdbscan>=0.8.29
elasticsearch==8.4.0
pickle5
faiss-cpu==1.7.4
pyvis
matplotlib>=3.7.1
networkx>=3.1
tqdm>=4.66.1
scikit-learn>=1.3.0
scipy>=1.11.4
statsmodels>=0.14.0
networkx>=3.1
spacy>=3.6.0
nltk>=3.8.1
fastapi >=0.95.2
uvicorn >=0.22.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
qdrant_client>=1.0.0
redis>=4.6.0
neo4j>=5.10.0
neo4j-driver>=5.10.0
httpx>=0.27.0